Stripped down requirements for RTD build Refactored tool argparse from tool CLI Renamed a few CLI-related source files
